+++ aldrin/src/promise.rs
@@ -1,0 +1,95 @@
+use crate::core::Serialize;
+use crate::error::Error;
+use crate::low_level::Promise as LlPromise;
+use std::fmt;
+use std::marker::PhantomData;
+
+/// Replies to a pending call.
+pub struct Promise<T: ?Sized, E: ?Sized> {
+    inner: LlPromise,
+    phantom: PhantomData<fn(T, E)>,
+}
+
+impl<T: ?Sized, E: ?Sized> Promise<T, E> {
+    pub(crate) fn new(inner: LlPromise) -> Self {
+        Self {
+            inner,
+            phantom: PhantomData,
+        }
+    }
+
+    /// Casts the promise to a different result type.
+    pub fn cast<T2: ?Sized, E2: ?Sized>(self) -> Promise<T2, E2> {
+        Promise::new(self.inner)
+    }
+
+    /// Extracts the inner low-level promise.
+    pub fn into_low_level(self) -> crate::low_level::Promise {
+        self.inner
+    }
+
+    /// Aborts the call.
+    ///
+    /// The caller will be notified that the call was aborted.
+    pub fn abort(self) -> Result<(), Error> {
+        self.inner.abort()
+    }
+
+    /// Signals that an invalid function was called.
+    pub fn invalid_function(self) -> Result<(), Error> {
+        self.inner.invalid_function()
+    }
+
+    /// Signals that invalid arguments were passed to the function.
+    pub fn invalid_args(self) -> Result<(), Error> {
+        self.inner.invalid_args()
+    }
+}
+
+impl<T, E> Promise<T, E>
+where
+    T: Serialize + ?Sized,
+    E: ?Sized,
+{
+    /// Signals that the call was successful.
+    pub fn ok(self, value: &T) -> Result<(), Error> {
+        self.inner.ok(value)
+    }
+}
+
+impl<E: ?Sized> Promise<(), E> {
+    /// Signals that the call was successful without returning a value.
+    pub fn done(self) -> Result<(), Error> {
+        self.inner.done()
+    }
+}
+
+impl<T, E> Promise<T, E>
+where
+    T: ?Sized,
+    E: Serialize + ?Sized,
+{
+    /// Signals that the call failed.
+    pub fn err(self, value: &E) -> Result<(), Error> {
+        self.inner.err(value)
+    }
+}
+
+impl<T, E> Promise<T, E>
+where
+    T: Serialize + ?Sized,
+    E: Serialize + ?Sized,
+{
+    /// Sets the call's reply.
+    pub fn set(self, res: Result<&T, &E>) -> Result<(), Error> {
+        self.inner.set(res)
+    }
+}
+
+impl<T: ?Sized, E: ?Sized> fmt::Debug for Promise<T, E> {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        f.debug_struct("Promise")
+            .field("inner", &self.inner)
+            .finish()
+    }
+}